refactor: segregate helper based on functionality
